3232745 - Avoid Losing Query Tags when Changing Display Override Status
This submit callback gets invoked twice when changing a display to make
it override the default display and vice-versa. Previously, the second
time this was invoked, the value set in the form state during the
previous call would be obliterated, leading to the query tags being
lost. Now, we bail out before this happens.

This needs tests.