layout_builder: always allow overrides storage creation
The overrides section storage plugin in Layout Builder previously
prohibited the creation of a plugin instance when the associated entity
view mode did not permit overrides, which negatively impacts local task
generation by forcing a max-age of 0 on entities without layout
overrides enabled.

This change to the overrides section storage allows plugin instances to
be created regardless of the entity view mode's ability to host such
overrides and delays the intended effect of disallowing overrides until
the access check on the local task's associated route is performed.

In addition to the change to the overrides section storage, two tests
needed to be adjusted to compensate for the change in behavior.

An additional test was added to ensure that Layout Builder's local tasks
no longer prevent pages from being cached.