2951134 - Fix Term Name Generation in ER Autocreate Test
Using `randomString()` to generate the names of terms can lead to
quote and comma characters being included in the randomly-generated test
data. When this happens, the test fails because the term name doesn't
parse as intended. This is now fixed by using the sentence generator
for this instead. As a bonus, we are also verifying that spaces do not
throw off the auto-create widget.