2732429: Also process fields with interface for files.
Currently this only works for fields which inherit from TextItemBase.
This requires custom modules to do quite a bit of refactoring to work.

Using the interface provides a simpler approach to retrieving formatted
text from a custom field.

This uses the fully qualified class name instead of imports to avoid
conflicting changes while this approach is being investigated.