Documented hook_condition_info_alter
This commit updates the documentation for the hook_condition_info_alter in Drupal

This information will be valuable for other developers who need to understand or make changes to the code in the future.

By documenting the hook_condition_info_alter, this commit contributes to a better understanding of the Drupal codebase and helps to ensure the longevity and maintainability of the project.