Do not load language config overrides for the default language
Configs exported for a site's default language has no associated
language. Currently LanguageConfigFactoryOverride::loadOverrides
attempts to load an override for it anyway.

This also ends up storing serialized FALSE values into cache
for the default language collection. It looks safe to skip the
process if we are dealing with the default language.