Fix a couple LanguageTestTrait return values
Those two methods were returning the save() method return, which is an
int, instead of the expected object.

By chance, on currently used calls the return of them is not used, so
there was no error around them not returning the related object.

Change them to be accurate with documentation, and actual intention.