Revert line number test fail changes
Locally, UncaughtExceptionTest passes without this change, but it fails
on gitlab CI on both testUncaughtFatalError() and testLoggerException().

Just out of curiosity let us see how this partly revert of the line
numbers affects the result if at all.

It may be some kind of code cache somewhere in the stack for gitlab CI,
so if this passes, that maybe the case, and there is infrastructure to
check.