2796341 - MRF: Extract Logic for Obtaining ER Handle and Add TODO for DI
After applying the patch for this issue, there were multiple places
where the entity-reference-selection handler was being obtained with
identical logic. The MR feedback was that these calls should be updated
to use DI, but that's beyond the scope of this issue because it would
require reworking how the form element plug-in gets instantiated.

Instead, I've moved the logic for obtaining the handler to its own
utility method and added a TODO to address the lack of DI as part of