Issue #3425212: Migrate BigPipe implementation to use Revolt Event Loop
This removes the manual Fiber juggling that was added to BigPipe and
instead uses a new helper function in the form of `stream` which
executes a set of synchronous operations on the Revolt EventLoop.

Because of BigPipe's dependency on the renderer and the assumptions the
Renderer makes about how it runs in fibers we must also already convert
the code for the renderer to use the event loop.