Issue #3426047: Bump PHPStan to level 9
This pushes PHPStan to the highest level so we accept new code with
great typings while updating the PHPStan baseline to make sure we don't
immediately need to fix all existing issues.

The `phpstan-baseline.php` is excluded from analysis because it doesn't
actually impact our code and is automatically generated by PHPStan.

Ignore iterable types for arrays in PHPStan

We have a lot of arrays for different usages in Drupal (such as render
arrays) and at the current point in time it'd be very consuming to type
them all. We also expect we'll need to introduce new arrays like this
which might be difficult to type without cleaning up other things. To
make life of contributors easier we ignore those types of errors for
now.

We don't ignore
[`checkMissingIterableValueType`](https://phpstan.org/config-reference#vague-typehints)
entirely because there are other iterable types being used which have
value in receiving typehints.

We also ensure the baseline doesn't get added to archives to reduce the
amount of data downloaded outside of development.